Better handling of created/updated time; sorting by created
gofeeds doesn't do any normalization of the created and updated
keys of its feeds; we add a simple parser that tries to handle
common values (RFC3339 and RFC1123). Additionally we are now
sorting the output items by created date (which will have more
obvious value as we start to mux different feeds).