Initial tests on the filter package; ignore whitespace
This changeset adds an initial unit test for filters, including
basic matching functionality, the magic "*" value when creating
a regexp filter, and demonstration of matching on multi-word
tokens.

Additionally, we now trim whitespace from the strings that we
transform into regexps for filtering, and there is a test confirming
this behavior.