Switching to pointers for ItemFilters; fixing race condition
Two changes here -- in order to allow mutability on our filters,
we're now passing them around as pointers. (This may change as I
experiment further.) Also, a stupid bug was fixed--we pass the URL and
filters into our goroutines when processing the feeds, to avoid race
conditions where we processed the same feed multiple times.