Fixing logic error in feed process; adding Count filter
We're now filtering using the same logic that we experience in
most functional languages when we talk about filters--an Item stays in
the output only if it *passes* each filter.

Additionally, we've written a new filter type, Count. This adds some
complexity, as it's our first (and hopefully only) non-idempotent
filter, so we had to rearrange some behaviors.